# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pyrogram String Session Generator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SessionStart {
  /** Telegram API ID from my.telegram.org */
  api_id: number;
  /** Telegram API Hash from my.telegram.org */
  api_hash: string;
  /** Phone number in international format (e.g. +1234567890) */
  phone_number: string;
}

export interface SessionStarted {
  /** Temporary session ID to use in subsequent requests */
  session_id: string;
  /** Hash returned by Telegram for OTP verification */
  phone_code_hash: string;
}

export interface SessionVerify {
  /** Session ID from startSession */
  session_id: string;
  /** OTP code received via Telegram */
  phone_code: string;
}

export interface Session2FA {
  /** Session ID from startSession */
  session_id: string;
  /** Two-factor authentication password */
  password: string;
}

export interface SessionResult {
  /** Whether 2FA password is required */
  needs_2fa: boolean;
  /**
     * The generated Pyrogram string session (null if needs_2fa is true)
     * @nullable
     */
  string_session?: string | null;
  /**
     * 2FA password hint if needs_2fa is true
     * @nullable
     */
  hint?: string | null;
}

export interface ErrorResponse {
  /** Error message */
  error: string;
}