# lib/api-zod/src/generated/types/session2FA.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pyrogram String Session Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface Session2FA {
  /** Session ID from startSession */
  session_id: string;
  /** Two-factor authentication password */
  password: string;
}