# lib/api-zod/src/generated/types/sessionResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pyrogram String Session Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface SessionResult {
  /** Whether 2FA password is required */
  needs_2fa: boolean;
  /**
     * The generated Pyrogram string session (null if needs_2fa is true)
     * @nullable
     */
  string_session?: string | null;
  /**
     * 2FA password hint if needs_2fa is true
     * @nullable
     */
  hint?: string | null;
}