# lib/api-zod/src/generated/types/sessionStart.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pyrogram String Session Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface SessionStart {
  /** Telegram API ID from my.telegram.org */
  api_id: number;
  /** Telegram API Hash from my.telegram.org */
  api_hash: string;
  /** Phone number in international format (e.g. +1234567890) */
  phone_number: string;
}