# lib/api-zod/src/generated/types/sessionStarted.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pyrogram String Session Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface SessionStarted {
  /** Temporary session ID to use in subsequent requests */
  session_id: string;
  /** Hash returned by Telegram for OTP verification */
  phone_code_hash: string;
}