# lib/api-zod/src/generated/types/sessionVerify.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pyrogram String Session Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface SessionVerify {
  /** Session ID from startSession */
  session_id: string;
  /** OTP code received via Telegram */
  phone_code: string;
}